requirements.txt
keras==2.15.0
numpy==1.24.3
Pillow==10.1.0
Pillow==10.1.0
Requests==2.31.0
streamlit==1.28.2
